requirements.txt
# FastAPI and Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.13.0

# Redis
redis[hiredis]==5.0.1
aioredis==2.0.1

# AI/ML
boto3==1.34.0
openai==1.6.1
sentence-transformers==2.2.2

# Vector Database
qdrant-client==1.7.0

# Temporal
temporalio==1.5.0

# NATS
nats-py==2.6.0

# HTTP Client
httpx==0.25.2
aiofiles==23.2.1

# Validation
email-validator==2.1.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment
python-dotenv==1.0.0

# Monitoring
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1